+++ tanks-game/emptylevel.h
@@ -4,7 +4,7 @@
 class EmptyLevel : public Level {
 public:
 	std::string getName() { return "empty"; }
-	std::unordered_map<std::string, double> getWeights();
+	std::unordered_map<std::string, float> getWeights();
 	void initialize();
 	
 	EmptyLevel();
